Refactoring via sonalint
