+++ src/main/java/umc/kkijuk/server/recruit/controller/response/ValidRecruitList.java
@@ -1,6 +1,5 @@
 package umc.kkijuk.server.recruit.controller.response;
 
-import lombok.Builder;
 import lombok.Getter;
 import umc.kkijuk.server.recruit.domain.ValidRecruitDto;
