+++ src/main/java/umc/kkijuk/server/review/infrastructure/ReviewJpaRepository.java
@@ -3,7 +3,6 @@ package umc.kkijuk.server.review.infrastructure;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import umc.kkijuk.server.review.domain.RecruitReviewDtoInterface;
 
 import java.util.List;
